pyproject.toml
[project]
name = "introproject"
version = "0.1.0"
description = "A BioPython-based project for genome and protein sequence analysis."
authors = []
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.81",  # BioPython for sequence parsing
    "argparse",  # CLI argument handling (built-in but sometimes needed explicitly)
    "torch (>=2.6.0,<3.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "einops (>=0.8.1,<0.9.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "xgboost (>=2.1.4,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.1"
pytest = "^8.3.4"
isort = "^6.0.1"

[tool.black]
line-length = 88 
target-version = ['py38'] 

[tool.flake8]
max-line-length = 88 
